test(mobile): refactor de la suite a centavos + tests de migración v14
Sprint flutter-integer-cents-v1, fase 5 de 5.

**Refactor de la suite existente**: 881 literales monetarios ×100 en
seeds y aserciones, más los `formatCents(N)` de los widget tests. La
conversión se hizo con allowlist de nombres de parámetro monetarios
(`amount`, `principalAmount`, `creditLimit`, `spent`, …) para no tocar
`paymentDay`, `sortOrder`, `limit`, `monthsRequested` ni otros enteros
no-monetarios. Las aserciones se ajustaron desde el output real del
runner, iterando hasta verde: cada falla era exactamente
`expected × 100 == actual`, lo que confirma que el refactor no cambió
semántica, sólo unidad.

Casos que necesitaron lectura de contexto y no reemplazo automático:
predicados dentro de `expectLater`, listas de montos esperados, INSERTs
crudos vía `customStatement` que bypasean el DAO, y el marcador
`"version": 2` del backup (ahora 3).

**Tests nuevos de migración** (`database_migration_test.dart`):
- MG-IC-01: las 9 columnas monetarias quedan declaradas INTEGER y los 3
  ratios siguen REAL.
- MG-IC-02: la conversión 13→14 es idempotente — re-ejecutarla sobre una
  BD ya migrada no vuelve a multiplicar por 100.
- MG-IC-03: las rutas defensivas 5/8/11/12/13 → 14 dejan las 9 columnas
  en INTEGER sin lanzar.
- MG-IC-04: **regresión del bug que originó el sprint** — con 10 cargos
  de 10/20 centavos, pagar exactamente la deuda derivada ya no dispara
  `overpay_debt` y la tarjeta queda en cero sin residuo.

**`amount_input_formatter_test`** reescrito para el contrato nuevo,
incluyendo round-trip parse → format y el rechazo de >2 decimales.

`lib/widgets/amount_formatter.dart` eliminado (reemplazado por
`lib/utils/money.dart`).

896 tests verdes. flutter analyze 0 errores (3 hints preexistentes).

+++ mobile/test/helpers/factories.dart
@@ -25,7 +25,7 @@ class Factories {
   static AccountsCompanion credit({
     String? id,
     String name = 'Visa',
-    double creditLimit = 50000,
+    int creditLimit = 50000,
     int closingDay = 15,
     int paymentDay = 5,
   }) {
